feat: Implement movie search API with schedule information
