+++ build.gradle
@@ -1,21 +1,29 @@
 buildscript {
     ext {
-        queryDslVersion = "5.0.0"
+        springBootVersion = '3.2.2'
+        querydslVersion = '5.0.0'
     }
 }
 
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '3.1.0'
-    id 'io.spring.dependency-management' version '1.1.0'
-    id 'com.ewerk.gradle.plugins.querydsl' version '1.0.10'
+    id 'org.springframework.boot' version '3.2.2'
+    id 'io.spring.dependency-management' version '1.1.4'
+}
+
+bootJar {
+    enabled = false
+}
+
+jar {
+    enabled = true
 }
 
 allprojects {
     group = 'com.movie'
     version = '0.0.1-SNAPSHOT'
     sourceCompatibility = '17'
-
+    
     repositories {
         mavenCentral()
     }
@@ -25,45 +33,25 @@ subprojects {
     apply plugin: 'java'
     apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
-    apply plugin: 'com.ewerk.gradle.plugins.querydsl'
-
-    dependencies {
-        implementation 'com.querydsl:querydsl-jpa:${queryDslVersion}:jakarta'
-        annotationProcessor "com.querydsl:querydsl-apt:${queryDslVersion}:jakarta"
-        annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-        annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+    
+    java {
+        sourceCompatibility = '17'
     }
-
-    def querydslDir = "$buildDir/generated/querydsl"
-
-    querydsl {
-        jpa = true
-        querydslSourcesDir = querydslDir
-    }
-
-    sourceSets {
-        main.java.srcDir querydslDir
-    }
-
+    
     configurations {
-        querydsl.extendsFrom compileClasspath
-    }
-
-    compileQuerydsl {
-        options.annotationProcessorPath = configurations.querydsl
-    }
-
-    repositories {
-        mavenCentral()
-    }
-
-    dependencies {
-        testImplementation 'org.junit.jupiter:junit-jupiter:5.9.2'
-    }
-
-    tasks.withType(Test).tap {
-        configureEach {
-            useJUnitPlatform()
+        compileOnly {
+            extendsFrom annotationProcessor
         }
+    }
+    
+    dependencies {
+        implementation 'org.springframework.boot:spring-boot-starter'
+        implementation 'org.projectlombok:lombok:1.18.30'
+        annotationProcessor 'org.projectlombok:lombok:1.18.30'
+        testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    }
+    
+    test {
+        useJUnitPlatform()
     }
 }
