+++ docker-compose.yml
@@ -1,34 +1,39 @@
-version: '3.8'
+version: '3'
 
 services:
   mysql:
-    image: mysql:8.0
+    image: mysql:8.0.40
+    command: --character-set-server=utf8mb4 --collation-server=utf8mb4_unicode_ci --default-authentication-plugin=mysql_native_password
     environment:
-      MYSQL_ROOT_PASSWORD: rootpassword
+      MYSQL_ROOT_PASSWORD: root
       MYSQL_DATABASE: moviedb
-      MYSQL_USER: movieuser
-      MYSQL_PASSWORD: moviepassword
+      MYSQL_ROOT_HOST: '%'
     ports:
       - "3306:3306"
     volumes:
       - mysql_data:/var/lib/mysql
-    networks:
-      - app-network
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost"]
+      interval: 5s
+      timeout: 5s
+      retries: 5
 
   movie-api:
-    build: 
-      context: ./movie-api
-      dockerfile: Dockerfile
+    build:
+      context: .
+      dockerfile: api/Dockerfile
     ports:
       - "8080:8080"
-    depends_on:
-      - mysql
     environment:
-      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3306/moviedb
-      SPRING_DATASOURCE_USERNAME: movieuser
-      SPRING_DATASOURCE_PASSWORD: moviepassword
-    networks:
-      - app-network
+      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3306/moviedb?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC&characterEncoding=UTF-8
+      SPRING_DATASOURCE_USERNAME: root
+      SPRING_DATASOURCE_PASSWORD: root
+      SPRING_JPA_HIBERNATE_DDL_AUTO: update
+      SPRING_JPA_SHOW_SQL: 'true'
+      SPRING_JPA_PROPERTIES_HIBERNATE_FORMAT_SQL: 'true'
+    depends_on:
+      mysql:
+        condition: service_healthy
 
 volumes:
   mysql_data:
