+++ domain/src/main/java/com/movie/domain/entity/BaseEntity.java
