+++ domain/src/main/java/com/movie/domain/entity/Schedule.java
