+++ domain/src/main/java/com/movie/domain/entity/Seat.java
