+++ domain/src/main/java/com/movie/domain/entity/Theater.java
