+++ domain/src/main/java/com/movie/domain/repository/SeatRepository.java
