+++ domain/src/main/java/com/movie/domain/repository/TheaterRepository.java
