+++ infra/src/main/java/com/movie/infra/repository/SeatJpaRepository.java
