--- movie-api/Dockerfile
@@ -1,9 +1,0 @@
-FROM openjdk:17-jdk-slim
-
-WORKDIR /app
-
-COPY build/libs/*.jar app.jar
-
-EXPOSE 8080
-
-ENTRYPOINT ["java", "-jar", "app.jar"]
