--- movie-api/build.gradle
@@ -1,43 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.1.0'
-    id 'io.spring.dependency-management' version '1.1.0'
-}
-
-group = 'com.movie'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '17'
-
-compileJava {
-    options.compilerArgs += ['-parameters']
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(":movie-services")
-    implementation project(':movie-domain')
-    implementation project(':movie-infra')
-    
-    implementation 'org.springframework.boot:spring-boot-starter'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'com.h2database:h2'
-    
-    implementation 'org.projectlombok:lombok:1.18.30'
-    annotationProcessor 'org.projectlombok:lombok:1.18.30'
-    
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    
-    runtimeOnly 'com.mysql:mysql-connector-j:8.2.0'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}springBoot {
-    mainClass = 'com.movie.api.ApiApplication'
-}
-
