--- movie-api/src/main/resources/application.yml
@@ -1,24 +1,0 @@
-server:
-  port: 8080
-  servlet:
-    encoding:
-      charset: UTF-8
-      force: true
-
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://mysql:3306/moviedb
-    username: movieuser
-    password: moviepassword
-  jpa:
-    open-in-view: false
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQLDialect
-        format_sql: true
-    show-sql: true
-  main:
-    allow-bean-definition-overriding: true
