--- movie-domain/src/main/java/com/movie/domain/dto/MovieProjection.java
@@ -1,9 +1,0 @@
-package com.movie.domain.dto;
-
-public interface MovieProjection {
-    Long getId();
-    String getTitle();
-    String getThumbnail();
-    Integer getRunningTime();
-    String getGenre();
-}
