--- movie-infra/build.gradle
@@ -1,21 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.1.0' apply false
-    id 'io.spring.dependency-management' version '1.1.0'
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(":movie-domain")
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.1.0'
-    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
-
-    implementation 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
-
-    runtimeOnly 'com.mysql:mysql-connector-j:8.2.0'
-}
