--- movie-services/build.gradle
@@ -1,16 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.1.0' apply false
-    id 'io.spring.dependency-management' version '1.1.0'
-}
-
-dependencies {
-    implementation project(":movie-domain")
-    implementation project(":movie-infra")
-
-    implementation 'org.springframework.boot:spring-boot-starter:3.1.0'
-    implementation 'org.springframework:spring-tx:5.3.10'
-
-    implementation 'org.projectlombok:lombok:1.18.30'
-    annotationProcessor 'org.projectlombok:lombok:1.18.30'
-}
