+++ schema.sql
@@ -1,0 +1,47 @@
+CREATE TABLE IF NOT EXISTS movie (
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    title VARCHAR(255) NOT NULL,
+    grade VARCHAR(50),
+    genre VARCHAR(50),
+    running_time INTEGER,
+    release_date DATE,
+    thumbnail_url VARCHAR(255),
+    created_by VARCHAR(255),
+    created_at DATETIME,
+    updated_by VARCHAR(255),
+    updated_at DATETIME
+);
+
+CREATE TABLE IF NOT EXISTS theater (
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    name VARCHAR(255) NOT NULL,
+    created_by VARCHAR(255),
+    created_at DATETIME,
+    updated_by VARCHAR(255),
+    updated_at DATETIME
+);
+
+CREATE TABLE IF NOT EXISTS seat (
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    seat_number VARCHAR(10) NOT NULL,
+    theater_id BIGINT,
+    created_by VARCHAR(255),
+    created_at DATETIME,
+    updated_by VARCHAR(255),
+    updated_at DATETIME,
+    FOREIGN KEY (theater_id) REFERENCES theater(id)
+);
+
+CREATE TABLE IF NOT EXISTS schedule (
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    movie_id BIGINT,
+    theater_id BIGINT,
+    start_at DATETIME,
+    end_at DATETIME,
+    created_by VARCHAR(255),
+    created_at DATETIME,
+    updated_by VARCHAR(255),
+    updated_at DATETIME,
+    FOREIGN KEY (movie_id) REFERENCES movie(id),
+    FOREIGN KEY (theater_id) REFERENCES theater(id)
+);
