+++ services/build.gradle
@@ -1,0 +1,23 @@
+plugins {
+    id 'java'
+}
+
+dependencies {
+    implementation project(":domain")
+    implementation project(":infra")
+
+    implementation 'org.springframework.boot:spring-boot-starter'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework:spring-tx'
+    
+    implementation 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
+}
+
+bootJar {
+    enabled = false
+}
+
+jar {
+    enabled = true
+}
